cleanup

+++ drivers/soc/qcom/msm_bus/msm_bus_noc_adhoc.c
@@ -1,13 +1,6 @@
-/* Copyright (c) 2016, 2018, The Linux Foundation. All rights reserved.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2 and
- * only version 2 as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+// SPDX-License-Identifier: GPL-2.0-only
+/*
+ * Copyright (c) 2016, 2018, The Linux Foundation. All rights reserved.
  */
 
 #define pr_fmt(fmt) "AXI: NOC: %s(): " fmt, __func__
@@ -100,7 +93,7 @@ static uint64_t noc_bw(uint32_t bw_field, uint32_t qos_freq)
 /**
  * Calculate the max BW in Bytes/s for a given time-base.
  */
-static uint32_t noc_bw_ceil(long int bw_field, uint32_t qos_freq_khz)
+static uint32_t noc_bw_ceil(int bw_field, uint32_t qos_freq_khz)
 {
 	uint64_t bw_temp = 2 * qos_freq_khz * bw_field;
 	uint32_t scale = 1000 * BW_SCALE;
